refactor(memory): carry raw words across MMIO

+++ machine/cpp/machine/devices/audio/service_clock.h
@@ -44,11 +44,11 @@ public:
 	[[nodiscard]] auto sampleTransferStatusBits() const -> u32;
 
 private:
-	static u64 readTransferDataThunk(void* context, u32 addr, MappedBusSignals busSignals);
-	static void writeTransferAddressThunk(void* context, u32 addr, u64 value, MappedBusSignals busSignals);
-	static void writeTransferDataThunk(void* context, u32 addr, u64 value, MappedBusSignals busSignals);
+	static u32 readTransferDataThunk(void* context, u32 addr, MappedBusSignals busSignals);
+	static void writeTransferAddressThunk(void* context, u32 addr, u32 value, MappedBusSignals busSignals);
+	static void writeTransferDataThunk(void* context, u32 addr, u32 value, MappedBusSignals busSignals);
 	static bool transferDataWriteReadyThunk(void* context, u32 addr, MappedBusSignals busSignals);
-	static void writeTransferControlThunk(void* context, u32 addr, u64 value, MappedBusSignals busSignals);
+	static void writeTransferControlThunk(void* context, u32 addr, u32 value, MappedBusSignals busSignals);
 
 	void synchronizeBeforeTransferAccess(i64 nowCycles);
 	void advanceVoicesTo(i64 cycle);
